refactor: LottoTicket 의 toString 재정의 및 패키지 변경
